moved OkapiScriptEntryPointController and OkapiUrls to toplevel
